improved logic for saving and loading

+++ deval/base/validator.py
@@ -331,10 +331,19 @@ class BaseValidatorNeuron(BaseNeuron):
                 "start_over": self.start_over,
                 "queried_uids": self.queried_uids,
                 "hotkeys": self.hotkeys,
-                "past_weights": self.weights
             },
             os.path.join(save_path, "state.pt"),
         )
+
+        if save_weights:
+            torch.save(
+                {
+                    "past_weights": self.weights,
+                    "save_time": datetime.now()
+                },
+                os.path.join(save_path, "weights.pt"),
+            )
+
 
     def load_state(self):
         """Loads the state of the validator from a file."""
@@ -360,6 +369,7 @@ class BaseValidatorNeuron(BaseNeuron):
         self.hotkeys = state["hotkeys"]
         self.weights = state.get("past_weights", [])
 
+        # load historical contest and task repository
         try:
 
             with open(contest_path, "rb") as f: 
@@ -380,6 +390,21 @@ class BaseValidatorNeuron(BaseNeuron):
             bt.logging.warning(f"Unable to load the task repository or contest state, restarting contest: {e}")
             self.start_over = True
 
+        # try loading in historical weights
+        try:
+            weight_path = os.path.join(load_path, "weights.pt")
+            past_weights = torch.load(weight_path)
+            weight_save_time = past_weights.get("save_time")
+            if (datetime.now() - timedelta(hours=48)) <= weight_save_time:
+                self.weights = past_weights.get("past_weights", [])
+            else:
+                self.weights = []
+        
+        except Exception as e:
+            bt.logging.warning(f"Unable to load weights data with error: {e}")
+            self.weights = []
+
+
     def reset(self):
         self.weights = []
         self.task_repo = None
@@ -390,6 +415,12 @@ class BaseValidatorNeuron(BaseNeuron):
         files = os.listdir(load_path)
         for f in files:
             file_path = os.path.join(load_path, f)
+
+            # we want to maintain weights data over epochs 
+            if "weights.pt" in f:
+                continue
+
+            # otherwise we delete all save files 
             if os.path.isfile(file_path):
                 os.remove(file_path)
